# projects/com.gitlab.dekaulitz/src/controllers/controller_store.go
/*
 * Swagger Petstore
 *
 * This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.3
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package controllers

import (
	"net/http"
    "github.com/gin-gonic/gin"
    "com.gitlab.dekaulitz/src/vmodels"
)
/**
 * Delete purchase order by ID
 * 
 */
func DeleteOrder(c *gin.Context) {
    var orderId int64=c.Param("orderId")
    c.Writer.WriteHeader(http.StatusOK)
    return
}
/**
 * Returns pet inventories by status
 * 
 */
func GetInventory(c *gin.Context) {
    c.Writer.Header().Set("Content-Type","application/json")
    c.Writer.WriteHeader(http.StatusOK)
    var responseBody map[string]int32
    res,_:=json.Marshal(responseBody)
    c.Writer.Write(res)
    return
}
/**
 * Find purchase order by ID
 * 
 */
func GetOrderById(c *gin.Context) {
    var orderId int64=c.Param("orderId")
    c.Writer.Header().Set("Content-Type","application/json")
    c.Writer.Header().Set("Content-Type","application/xml")
    c.Writer.WriteHeader(http.StatusOK)
    var responseBody Order
    res,_:=json.Marshal(responseBody)
    c.Writer.Write(res)
    return
}
/**
 * Place an order for a pet
 * 
 */
func PlaceOrder(c *gin.Context) {
    var body vmodels.Order
    jsonFail := c.BindJSON(&body)
    if jsonFail != nil {
        panic(jsonFail)
    }
    c.Writer.Header().Set("Content-Type","application/json")
    c.Writer.Header().Set("Content-Type","application/xml")
    c.Writer.WriteHeader(http.StatusOK)
    var responseBody Order
    res,_:=json.Marshal(responseBody)
    c.Writer.Write(res)
    return
}